add_no_entry_message_fragment

+++ src/Controller/PostController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Post;
 use App\Form\PostType;
 use App\Repository\PostRepository;
+use App\Security\Voter\PostVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -12,15 +13,16 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Knp\Component\Pager\PaginatorInterface;
 
-
 #[Route('/post')]
 class PostController extends AbstractController
 {
     #[Route('/', name: 'app_post_index', methods: ['GET'])]
-    public function index(PostRepository $postRepository, Request $request, PaginatorInterface $paginator): Response
+    public function index(PostRepository $post, Request $request, PaginatorInterface $paginator): Response
     {
+        $this->denyAccessUnlessGranted(PostVoter::VIEW, $post);
+
         $pagination = $paginator->paginate(
-            $postRepository->paginationQuery(),
+            $post->paginationQuery(),
             $request->query->getInt('page', 1),
             10
         );
@@ -31,7 +33,7 @@ class PostController extends AbstractController
 
     #[Route('/new', name: 'app_post_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
-    {
+    {   
         $post = new Post();
         $post->setPublishedAt(new \DateTimeImmutable());
         $form = $this->createForm(PostType::class, $post);
@@ -58,6 +60,8 @@ class PostController extends AbstractController
     #[Route('/{id}', name: 'app_post_show', methods: ['GET'])]
     public function show(Post $post): Response
     {
+        $this->denyAccessUnlessGranted(PostVoter::VIEW, $post);
+
         return $this->render('post/show.html.twig', [
             'post' => $post,
         ]);
@@ -66,6 +70,8 @@ class PostController extends AbstractController
     #[Route('/{id}/edit', name: 'app_post_edit', methods: ['GET','POST','PUT'])]
     public function edit(Request $request, Post $post, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('POST_EDIT', $post);
+
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
 
@@ -90,6 +96,8 @@ class PostController extends AbstractController
     #[Route('/{id}', name: 'app_post_delete', methods: ['POST'])]
     public function delete(Request $request, Post $post, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(PostVoter::DELETE, $post);
+
         if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
             $entityManager->remove($post);
             $entityManager->flush();
